+++ packages/nextjs/src/utils/utils.test.ts
@@ -5,7 +5,7 @@ import { getEditingSecret } from './utils';
 describe('utils', () => {
   describe('getEditingSecret', () => {
     after(() => {
-      delete process.env.JSS_EDITING_SECRET;
+      delete process.env.SITECORE_EDITING_SECRET;
     });
 
     it('should throw if env variable missing', () => {
@@ -14,7 +14,7 @@ describe('utils', () => {
 
     it('should return env variable', () => {
       const secret = '1234abcd';
-      process.env.JSS_EDITING_SECRET = secret;
+      process.env.SITECORE_EDITING_SECRET = secret;
       const result = getEditingSecret();
       expect(result).to.equal(secret);
     });